Add files via upload
I attempted to put my code into a WPF format, but was struggling with time due to having two assignments the day after each other as well as an exam the week before. 

I added the ability to add recipe ingredients and save them, but didn't manage to get any further than that.